requirements.txt
aiohttp==3.8.4
aiomysql==0.1.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
databases==0.6.2
dataclasses-json==0.5.7
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
logtail-python==0.2.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.0
pydantic==1.10.4
PyMySQL==1.0.3
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
SQLAlchemy==1.4.41
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
